requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
apturl==0.5.2
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
awscli==1.22.34
beautifulsoup4==4.10.0
blinker==1.4
botocore==1.23.34
Brlapi==0.8.3
Brotli==1.0.9
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
cassandra-driver==3.27.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chrome-gnome-shell==0.0.0
cleo==2.0.1
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
crashtest==0.4.1
cryptography==40.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
docutils==0.17.1
duckdb==0.8.0
dulwich==0.21.5
fastapi==0.95.1
filelock==3.12.0
filetype==1.2.0
flatbuffers==23.5.9
frozenlist==1.3.3
geomet==0.2.1.post1
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
humanfriendly==10.0
idna==3.3
ImageHash==4.3.1
imageio==2.28.1
importlib-metadata==6.6.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.7.1
jmespath==0.10.0
jsonschema==4.17.3
kazam==1.4.5
keyring==23.13.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
llvmlite==0.40.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mutagen==1.45.1
netifaces==0.11.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.0
ObjectExtractor==0.2.2
olefile==0.46
onnxruntime==1.14.1
onnxruntime-gpu==1.14.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pexpect==4.8.0
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.1
poetry==1.5.0
poetry-core==1.6.0
poetry-plugin-export==1.3.1
pooch==1.7.0
proto-plus==1.22.2
protobuf==4.22.3
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==1.10.7
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMatting==1.1.8
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-multipart==0.0.6
pytz==2022.1
PyWavelets==1.4.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
rapidfuzz==2.15.1
rembg==2.0.36
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==1.0.0
roman==3.3
rsa==4.9
s3transfer==0.5.0
scikit-image==0.20.0
scipy==1.10.1
scour==0.38.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
starlette==0.26.1
sympy==1.12
systemd-python==234
tifffile==2023.4.12
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
trove-classifiers==2023.5.2
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.22.0
virtualenv==20.23.0
wadllib==1.3.6
watchdog==3.0.0
webencodings==0.5.1
websockets==9.1
xdg==5
xkit==0.0.0
yarl==1.9.2
yt-dlp==2022.4.8
zipp==1.0.0
